requirements.txt
asyncio >= 3.4.3
python-kasa >= 0.5.0
